+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * Drop support for Python 2.7, 3.4, and 3.5.
+* Correctly parse relative includes without jinja root.
 
 
 1.1.3 (2019-12-26)